Feat: support to sync the revision and the workflow status

+++ test/e2e-apiserver-test/suite_test.go
@@ -22,6 +22,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+	"os"
 	"strings"
 	"testing"
 	"time"
@@ -34,6 +35,7 @@ import (
 
 	"github.com/oam-dev/kubevela/pkg/apiserver"
 	"github.com/oam-dev/kubevela/pkg/apiserver/config"
+	"github.com/oam-dev/kubevela/pkg/apiserver/domain/service"
 	"github.com/oam-dev/kubevela/pkg/apiserver/infrastructure/clients"
 	"github.com/oam-dev/kubevela/pkg/apiserver/infrastructure/datastore"
 	apisv1 "github.com/oam-dev/kubevela/pkg/apiserver/interfaces/api/dto/v1"
@@ -82,9 +84,13 @@ var _ = BeforeSuite(func() {
 	By("wait for api server to start")
 	Eventually(
 		func() error {
+			password := os.Getenv("VELA_UX_PASSWORD")
+			if password == "" {
+				password = service.InitAdminPassword
+			}
 			var req = apisv1.LoginRequest{
 				Username: "admin",
-				Password: "VelaUX12345",
+				Password: password,
 			}
 			bodyByte, err := json.Marshal(req)
 			Expect(err).Should(BeNil())
@@ -118,11 +124,13 @@ var _ = BeforeSuite(func() {
 var _ = AfterSuite(func() {
 	By("tearing down the test environment")
 	var nsList v1.NamespaceList
-	err := k8sClient.List(context.TODO(), &nsList)
-	Expect(err).ToNot(HaveOccurred())
-	for _, ns := range nsList.Items {
-		if strings.HasPrefix(ns.Name, testNSprefix) {
-			_ = k8sClient.Delete(context.TODO(), &ns)
+	if k8sClient != nil {
+		err := k8sClient.List(context.TODO(), &nsList)
+		Expect(err).ToNot(HaveOccurred())
+		for _, ns := range nsList.Items {
+			if strings.HasPrefix(ns.Name, testNSprefix) {
+				_ = k8sClient.Delete(context.TODO(), &ns)
+			}
 		}
 	}
 })
